requirements.txt
Django
django-imagekit
django-crispy-forms
crispy-bootstrap5
python-decouple